# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FlowPay API - India's stablecoin payment super-layer
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /creator/products:
    get:
      operationId: listCreatorProducts
      tags: [creator]
      summary: List creator products
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CreatorProduct"
    post:
      operationId: createCreatorProduct
      tags: [creator]
      summary: Create a creator product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCreatorProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatorProduct"
  /creator/products/{id}:
    get:
      operationId: getCreatorProduct
      tags: [creator]
      summary: Get a creator product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatorProduct"
        "404":
          description: Not found
  /creator/products/{id}/purchase:
    post:
      operationId: purchaseCreatorProduct
      tags: [creator]
      summary: Purchase a creator product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseProductBody"
      responses:
        "201":
          description: Purchase complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatorSale"
...
  /webhooks/dodo:
    post:
      operationId: dodoWebhook
      tags: [creator]
      summary: Dodo Payments webhook handler — updates sale payment status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DodoWebhookBody"
      responses:
        "200":
          description: Webhook processed
        "400":
          description: Bad request

  # AI AGENTS
...
components:
  schemas:
...
    CreatorProduct:
      type: object
      properties:
        id:
          type: integer
        creatorName:
          type: string
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [course, newsletter, ebook, template, other]
        priceUsdg:
          type: string
        salesCount:
          type: integer
        totalRevenue:
          type: string
        isActive:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, creatorName, title, description, type, priceUsdg, salesCount, totalRevenue, isActive, createdAt]
    CreateCreatorProductBody:
      type: object
      properties:
        creatorName:
          type: string
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [course, newsletter, ebook, template, other]
        priceUsdg:
          type: string
      required: [creatorName, title, description, type, priceUsdg]
    PurchaseProductBody:
      type: object
      properties:
        buyerEmail:
          type: string
        buyerName:
          type: string
      required: [buyerEmail, buyerName]
    CreatorSale:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productTitle:
          type: string
        creatorName:
          type: string
        buyerName:
          type: string
        buyerEmail:
          type: string
        amountUsdg:
          type: string
        feeUsdg:
          type: string
        creatorReceives:
          type: string
        solanaSignature:
          type: ["string", "null"]
        dodoSessionId:
          type: ["string", "null"]
        dodoCheckoutUrl:
          type: ["string", "null"]
        dodoPaymentStatus:
          type: string
          enum: [pending, paid, failed]
        createdAt:
          type: string
          format: date-time
      required: [id, productId, productTitle, creatorName, buyerName, buyerEmail, amountUsdg, feeUsdg, creatorReceives, dodoPaymentStatus, createdAt]
    DodoWebhookBody:
      type: object
      description: Dodo Payments webhook event in the standardwebhooks format
      properties:
        type:
          type: string
          description: Event type, e.g. payment.succeeded or payment.failed
        data:
          type: object
          description: Event payload — for payment events contains checkout_session_id and payment_id
          properties:
            checkout_session_id:
              type: string
            payment_id:
              type: string
          additionalProperties: true
      required: [type, data]
...